django/blog/requirements.txt
Django==4.0.5
django_ckeditor==6.4.2
python-dotenv==0.20.0
